# shared/api_spec.yaml
openapi: 3.0.3
info:
  title: DPS Dashboard API
  description: >
    Deshapriya Park Durga Puja Club Management Dashboard API.
    All endpoints require authentication unless marked as public.
    Role enforcement is server-side on every route.
  version: 1.0.0
  contact:
    name: DPS Dashboard Team
servers:
  - url: http://localhost:3000/api
    description: Local development
  - url: http://{vps_ip}/api
    description: LunaNode VPS
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: next-auth.session-token
      description: HTTP-only JWT session cookie set by NextAuth
  schemas:
...
    MembershipStatus:
      type: string
      enum: [PENDING_APPROVAL, PENDING_PAYMENT, ACTIVE, EXPIRED, SUSPENDED]
    MembershipType:
      type: string
      enum: [MONTHLY, HALF_YEARLY, ANNUAL]
...
    ApprovalStatus:
      type: string
      enum: [PENDING, APPROVED, REJECTED]
    ApprovalEntityType:
      type: string
      enum: [TRANSACTION, MEMBER_ADD, MEMBER_EDIT, MEMBER_DELETE, MEMBERSHIP]
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
        details:
          type: object
...
    MemberCreate:
      type: object
      required: [name, email, phone, address]
      properties:
        name:
          type: string
          maxLength: 255
        email:
          type: string
          format: email
        phone:
          type: string
          pattern: '^\+91\d{10}$'
        address:
          type: string
...
    MemberResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        memberId:
          type: string
          description: "Format: DPC-YYYY-NNNN-00"
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        address:
          type: string
        membershipStatus:
          $ref: '#/components/schemas/MembershipStatus'
        joinedAt:
          type: string
          format: date-time
        subMembers:
          type: array
          items:
            $ref: '#/components/schemas/SubMemberResponse'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    SubMemberResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        memberId:
          type: string
          description: "Format: DPC-YYYY-NNNN-SS (01-03)"
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        relation:
          type: string
        canLogin:
          type: boolean
        createdAt:
          type: string
          format: date-time

    # Memberships
    MembershipCreate:
      type: object
      required: [memberId, type]
      properties:
        memberId:
          type: string
          format: uuid
        type:
          $ref: '#/components/schemas/MembershipType'
        isApplicationFee:
          type: boolean
          default: false
...
    MembershipResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        memberId:
          type: string
          format: uuid
        type:
          $ref: '#/components/schemas/MembershipType'
        amount:
          type: number
          format: decimal
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        isApplicationFee:
          type: boolean
        status:
          $ref: '#/components/schemas/ApprovalStatus'
        createdAt:
          type: string
          format: date-time

    # Payments
    CreateOrderRequest:
      type: object
      required: [amount, membershipType]
      properties:
        amount:
          type: number
          format: decimal
          description: "Must match exact fee: 250 (monthly), 1500 (half-yearly), 3000 (annual), 10000 (application)"
        membershipType:
          $ref: '#/components/schemas/MembershipType'
        memberId:
          type: string
          format: uuid
          description: "Member paying for (required for pay-on-behalf)"
        isApplicationFee:
          type: boolean
          default: false
    CreateOrderResponse:
      type: object
      properties:
        orderId:
          type: string
        amount:
          type: number
        currency:
          type: string
          default: INR
        keyId:
          type: string
          description: "Razorpay key ID for client-side checkout"
    VerifyPaymentRequest:
      type: object
      required: [razorpay_order_id, razorpay_payment_id, razorpay_signature]
      properties:
        razorpay_order_id:
          type: string
        razorpay_payment_id:
          type: string
        razorpay_signature:
          type: string
    VerifyPaymentResponse:
      type: object
      properties:
        verified:
          type: boolean
        transactionId:
          type: string
          format: uuid
        receiptNumber:
          type: string

    # Transactions
...
    ApprovalResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        entityType:
          $ref: '#/components/schemas/ApprovalEntityType'
        entityId:
          type: string
          format: uuid
        action:
          type: string
        previousData:
          type: object
          nullable: true
        newData:
          type: object
          nullable: true
        requestedBy:
          type: object
          properties:
            id:
              type: string
            name:
              type: string
        status:
          $ref: '#/components/schemas/ApprovalStatus'
        reviewedBy:
          type: object
          nullable: true
          properties:
            id:
              type: string
            name:
              type: string
        reviewedAt:
          type: string
          format: date-time
          nullable: true
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
...
    PaginatedResponse:
      type: object
      properties:
        data:
          type: array
          items: {}
        total:
          type: integer
        page:
          type: integer
        pageSize:
          type: integer
        totalPages:
          type: integer
  parameters:
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        default: 1
        minimum: 1
    PageSizeParam:
      name: pageSize
      in: query
      schema:
        type: integer
        default: 20
        minimum: 1
        maximum: 100
    SearchParam:
      name: search
      in: query
      schema:
        type: string
    StatusFilterParam:
      name: status
      in: query
      schema:
        type: string
security:
  - cookieAuth: []
paths:
...
  /members:
    get:
      tags: [Members]
      summary: List all members
      description: "Admin and Operator only. Supports pagination, search, status filter."
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PageSizeParam'
        - $ref: '#/components/parameters/SearchParam'
        - $ref: '#/components/parameters/StatusFilterParam'
      responses:
        '200':
          description: Paginated list of members
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PaginatedResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/MemberResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Forbidden (MEMBER role)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    post:
      tags: [Members]
      summary: Create a new member
      description: >
        Admin: creates directly (auto-approved).
        Operator: creates Approval record (MEMBER_ADD), member not added until admin approves.
        Auto-generates DPC-YYYY-NNNN-00 member ID.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MemberCreate'
      responses:
        '201':
          description: Member created (or approval pending for operator)
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/MemberResponse'
                  - $ref: '#/components/schemas/ApprovalResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /memberships:
    get:
      tags: [Memberships]
      summary: List memberships
      description: "Admin/Operator: all. Member: own only."
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PageSizeParam'
        - name: memberId
          in: query
          schema:
            type: string
            format: uuid
        - $ref: '#/components/parameters/StatusFilterParam'
      responses:
        '200':
          description: Paginated memberships
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PaginatedResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/MembershipResponse'

    post:
      tags: [Memberships]
      summary: Create membership period
      description: >
        Creates a new membership period. Amount auto-calculated from type.
        Operator: creates Approval record (MEMBERSHIP).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MembershipCreate'
      responses:
        '201':
          description: Membership created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MembershipResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /payments/create-order:
    post:
      tags: [Payments]
      summary: Create Razorpay order
      description: >
        Creates a Razorpay order for UPI or bank transfer payment.
        Amount must match exact fee for selected membership type.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrderRequest'
      responses:
        '200':
          description: Razorpay order created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateOrderResponse'
        '400':
          description: Amount mismatch or validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /payments/verify:
    post:
      tags: [Payments]
      summary: Verify Razorpay payment signature
      description: "Client-side verification after checkout. Confirms payment with Razorpay."
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VerifyPaymentRequest'
      responses:
        '200':
          description: Payment verified
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VerifyPaymentResponse'
        '400':
          description: Signature verification failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # ==================== WEBHOOKS ====================
...